+++ src/style/chronograf.scss
@@ -137,9 +137,9 @@
 @import 'src/checkout/SuccessOverlay.scss';
 @import 'src/checkout/CheckoutPage.scss';
 @import 'src/checkout/CheckoutForm.scss';
+@import 'src/operator/LimitsField.scss';
 @import 'src/operator/OperatorNav.scss';
 @import 'src/operator/OrgOverlay.scss';
-@import 'src/operator/LimitsLabel.scss';
 @import 'src/operator/account/AccountGrid.scss';
 @import 'src/writeData/components/fileUploads/FileUploads.scss';
 @import 'src/secrets/components/secrets.scss';
